requirements.txt
pennylane
torch
torchvision
torchaudio
scikit-learn
pandas
matplotlib
scipy
seaborn
plotly
pyyaml
